dropped execvp() prototype mangle hack
Made main() and serve() take char ** argv (instead of const char **),
so that the system header defined execvp() prototype can be used
(char * const * mismatches const char * as latter makes the data
pointed to read-only). As data pointed by argv[] is in modifiable
memory it can be passed as is...

This change reduces compilation warnings with older compilers;
newer compilers (that support "'-Wbuiltin-declaration-mismatch')
were silent when this hack was user.